Utilize the module hierarchy to shorten template type names
