+++ src/def/grib2.rs
@@ -1,3 +1,4 @@
 pub use section5::*;
 
 mod section5;
+pub mod template5;
